add support for params in python; refactor to remove warnings

+++ src/segments.cpp
@@ -31,7 +31,7 @@ Segments::~Segments()
     delete[] segment;
     segment = nullptr;
     _area = 0;
-    for (int i = 0; i < seg_data.size(); i++) {
+    for (size_t i = 0; i < seg_data.size(); i++) {
         delete seg_data[i];
     }
 }
@@ -44,7 +44,7 @@ uint32_t Segments::area()
 void Segments::reset()
 {
     memset(segment, -1, sizeof(uint32_t) * _area);
-    for (int i = 0; i < seg_data.size(); i++) {
+    for (size_t i = 0; i < seg_data.size(); i++) {
         delete seg_data[i];
     }
     seg_data.clear();
